+++ benchmark/src/benchmarks/view-nest-instance.js
@@ -3,9 +3,7 @@
 export function create({ Nest }) {
 	return () => {
 		return Nest({
-			children: () => {
-				return () => "test";
-			},
+			children: () => "test",
 		});
 	};
 }